Fix package naming and style issues
