--- src/main/java/seedu/EquipmentMaster/exception/EquipmentMasterException.java
@@ -1,7 +1,0 @@
-package seedu.EquipmentMaster.exception;
-
-public class EquipmentMasterException extends Exception {
-    public EquipmentMasterException (String message) {
-        super(message);
-    }
-}
