+++ src/main/java/seedu/equipmentmaster/common/Messages.java
@@ -1,0 +1,9 @@
+package seedu.equipmentmaster.common;
+
+public class Messages {
+    public static final String MESSAGE_WELCOME = "Welcome to Equipment Master! How can I help you today?";
+    public static final String MESSAGE_GOODBYE = "Welcome to Equipment Master! How can I help you today?";
+    public static final String MESSAGE_INVALID_INPUT = "OOPS! Sorry I don't understand this :(";
+    public static final String MESSAGE_INVALID_ADD_FORMAT = "Please enter with the right format (with n/ and q/)";
+    public static final String MESSAGE_INVALID_FIND_FORMAT = "Please provide a keyword to find. Example: find STM32";
+}
