+++ src/main/java/seedu/equipmentmaster/exception/EquipmentMasterException.java
@@ -1,0 +1,7 @@
+package seedu.equipmentmaster.exception;
+
+public class EquipmentMasterException extends Exception {
+    public EquipmentMasterException (String message) {
+        super(message);
+    }
+}
